QraphQL added

+++ src/main/java/com/pluralsight/repository/ApplicationRepository.java
@@ -3,5 +3,5 @@ package com.pluralsight.repository;
 import com.pluralsight.entity.Application;
 import org.springframework.data.repository.CrudRepository;
 
-public interface ApplicationRepository extends CrudRepository<Application, Long> {
+public interface ApplicationRepository extends CrudRepository<Application, Integer> {
 }
